+++ R/get_solution.R
@@ -1,43 +1,86 @@
-#' Get an Orbital Solution
+#' Get an Astronomical Solution
 #'
-#' @param orbital_solution Character vector with the name of the orbital
-#'   solution to use. One of `"ZB18a"` (default) from Zeebe and Lourens (2019),
-#'   or `"La11"` (not yet implemented!).
-# quiet and force are documented in get_ZB18a
-#' @inheritParams get_ZB18a
-#' @seealso [get_ZB18a()]
-#' @inherit get_ZB18a references
-#' @returns `get_solution()` returns a [tibble][tibble::tibble-package] with the
-#'   orbital solution input and some preprocessed new columns.
+#' Download supported astronomical solutions from the web and store it in the
+#' user's cache directory. The next use of the function will load the data from
+#' the cache rather than downloading it again. This also provides a wrapper for
+#' [astrochron::getLaskar()] if one of their supported solutions is specified,
+#' but converts the output to a [tibble][tibble::tibble-package]. Note that we
+#' do not cache these solutions locally, however.
+#'
+# astronomical_solution, quiet, and force are documented in get_ZB
+#' @inheritParams get_ZB
+#' @inherit get_ZB references
+#' @seealso [PT_ZB18a], [ZB17], [ZB18a], [ZB20]
+#' @returns A [tibble][tibble::tibble-package] with the astronomical solution
+#'   (and some preprocessed new columns).
 #' @examples
 #' \donttest{
-#' get_solution()
+#' get_solution("PT-ZB18a")
+#' get_solution("ZB20a")
+#' get_solution("La11")
 #' }
 #' @export
-get_solution <- function(orbital_solution = "ZB18a", quiet = FALSE, force = FALSE) {
-  solutions <- c("ZB18a", "La11")
-  if (!orbital_solution %in% solutions) {
-    cli::cli_abort(c("{.var orbital_solution} must be one of: {.or {.q {solutions}}}",
-                     "x" = "You've supplied {.q {orbital_solution}}"))
+get_solution <- function(astronomical_solution = "PT-ZB18a", quiet = FALSE, force = FALSE) {
+  if ("data.frame" %in% class(astronomical_solution)) {
+    return(prepare_solution(astronomical_solution, quiet = quiet))
   }
 
-  if (orbital_solution == "ZB18a") {
-    # read in the (new?) cached result
-    dat <- get_ZB18a(quiet = quiet, force = force)
+  # effectively this is a wrapper that checks only for valid solutions...
+  solutions <- c("PT-ZB18a", # the default
+                 # special case to catch ambiguous naming
+                 "ZB18a",
+                 "PT-La11", # just to thow an error
+                 # we rely on astrochron to get these for us
+                 "La04", "La10a", "La10b", "La10c", "La10d", "La11",
+                 # we get these ourselves
+                 "ZB17a", "ZB17b", "ZB17c", "ZB17d",
+                 "ZB17e", "ZB17f", "ZB17h", "ZB17i",
+                 "ZB17j", "ZB17k", "ZB17p",
+                 "ZB18a-100",
+                 "ZB18a-300",
+                 "ZB20a", "ZB20b", "ZB20c", "ZB20d")
+
+  if (!astronomical_solution %in% solutions) {
+    cli::cli_abort(c("{.var astronomical_solution} must be one of: {.or {.q {solutions}}}",
+                     "x" = "You've supplied {.q {astronomical_solution}}"))
   }
-  if (orbital_solution == "La11") {
-    ## dat <- snvecR::La11
-    cli::cli_abort(c("Orbital solution: La11 currently not supported.",
-      "!" = "The input OS for snvec must be in the Heliocentric Inertial Reference frame (HCI) (J2000).",
-      "x" = "The La11 solution is in the invariant reference frame.",
-      "i" = "Pull requests welcome."))
+
+  if (astronomical_solution == "ZB18a") {
+    cli::cli_abort(c("i" = "There are multiple versions of {.q ZB18a} available.",
+                     "!" = "Please select the one you want:",
+                     "*" = "{.q PT-ZB18a}: for the precession-tilt solutions",
+                     "*" = "{.q ZB18a-100}: for the eccentricity and inclination for the past 100 Ma",
+"*" = "{.q ZB18a-300}: for the eccentricity and inclination for the past 300 Ma"))
+  }
+
+  if (astronomical_solution == "PT-La11") {
+    cli::cli_abort(c("i" = "Astronomical solution {.q PT-La11} is not supported.",
+                     "!" = "The input OS for snvec must be either in the:",
+                     "*" = "Heliocentric inertial reference frame (HCI)",
+                     "*" = "Ecliptic reference frame (J2000).",
+                     "x" = "The La11 solution is in the invariant/inertial reference frame.",
+                     "i" = "To resolve this, you need the positions/velocities and masses of all the bodies.",
+                     "i" = "Or the angles between their inertial reference frame and J2000.",
+                     "i" = "Pull requests welcome."))
+  }
+
+  if (grepl("^La[0-9][a-z]?", astronomical_solution)) {
+    dat <- astrochron::getLaskar(sol = tolower(astronomical_solution)) |>
+      tibble::as_tibble()
+  }
+
+  if (astronomical_solution == "PT-ZB18a" ||
+        grepl("^ZB[0-9][0-9][a-z]", astronomical_solution)) {
+    dat <- get_ZB(astronomical_solution, quiet = quiet, force = force)
   }
 
   return(dat)
 }
 
-#' Get the ZB18a solution
-#'
+# #' Get a ZB solution
+# #'
+#' @param astronomical_solution Character vector with the name of the desired
+#'   solution.
 #' @param quiet Be quiet?
 #'
 #'   * If `TRUE`, hide info messages.
@@ -45,54 +88,79 @@ get_solution <- function(orbital_solution = "ZB18a", quiet = FALSE, force = FALS
 #'   * If `FALSE` (the default) print info messages and timing.
 #' @param force Force re-downloading the results, even if the solution is saved
 #'   to the cache.
-#' @returns `get_ZB18a()` returns a [tibble][tibble::tibble-package] with the
-#'   orbital solution input and some preprocessed new columns.
-#' @rdname ZB18a
-#' @seealso [prepare_solution()] Processes orbital solution input to include
-#'   helper columns.
-#' @examples
-#' \donttest{
-#' get_ZB18a()
-#' }
-#' @export
-get_ZB18a <- function(quiet = FALSE, force = FALSE) {
-  ZB18a_url <- "http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat"
-
+#' @returns A [tibble][tibble::tibble-package] with the astronomical solution
+#'   input and, in the case of the PT-ZB18a, some preprocessed new columns.
+# #' @seealso [prepare_solution()] Processes precession-tilt astronomical
+# #'   solution input to include helper columns.
+#' @inherit PT_ZB18a references
+# #' @examples
+# #' \donttest{
+# #' get_ZB()
+# #' }
+# NOTE: I don't export this anymore.
+get_ZB <- function(astronomical_solution = "PT-ZB18a",
+                   quiet = FALSE,
+                   force = FALSE) {
+  base_url <- "http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/"
   cachedir <- tools::R_user_dir("snvecR", which = "cache")
 
-  raw_path <- paste0(cachedir, "/ems-plan3.dat")
-  csv_path <- paste0(cachedir, "/ZB18a.csv")
-  rds_path <- paste0(cachedir, "/ZB18a.rds")
+  if (astronomical_solution == "PT-ZB18a") {
+    url <- paste0(base_url, "PrecTilt/OS/ZB18a/ems-plan3.dat")
+    # TODO: rename to PrecTilt, same as on ZB's website
+    raw_col_names <- c("t", # time in days
+                       "aa", # semimajor axis
+                       "ee", # eccentricity
+                       "inc", # inclination
+                       "lph", # long periapse
+                       "lan", # long ascending node
+                       "arp", # argument of periapse
+                       "mna") # mean anomaly
+  } else {
+    raw_col_names <- c("age", # time in ka, negative
+                       "ecc", # eccentricity
+                       "inc") # inclination
+  }
+  # different URLs for different solutions
+  if (grepl("^ZB1[78][a-z](-100)?$", astronomical_solution)) {
+    # 17 and 18 are stored in the root directory
+    url <- glue::glue("{base_url}{gsub('-100', '', astronomical_solution)}.dat")
+  } else if (grepl("^ZB[0-9][0-9][a-z](-300)?$", astronomical_solution)){
+    # ZB20 is stored in 300Myr subdirectory
+    url <- glue::glue("{base_url}300Myr/{gsub('-300', '', astronomical_solution)}.dat")
+  }
 
-  raw_col_names <- c("t", # time in days
-                     "aa", # semimajor axis
-                     "ee", # eccentricity
-                     "inc", # inclination
-                     "lph", # long periapse
-                     "lan", # long ascending node
-                     "arp", # argument of periapse
-                     "mna") # mean anomaly
+  # where will we save our cached results?
+  raw_path <- paste0(cachedir, "/", astronomical_solution, ".dat")
+  csv_path <- paste0(cachedir, "/", astronomical_solution, ".csv")
+  rds_path <- paste0(cachedir, "/", astronomical_solution, ".rds")
 
   # read final processed file from cache if available
   if (!force && file.exists(rds_path)) {
-    ZB18a <- readr::read_rds(rds_path)
-    return(ZB18a)
+    rds <- readr::read_rds(rds_path)
+    return(rds)
   }
 
   # read raw intermediate steps from disk
   if (!force && (file.exists(csv_path) || file.exists(raw_path))) {
     if (file.exists(csv_path)) {
-      ZB18a_raw <- readr::read_csv(csv_path, show_col_types = FALSE)
+      raw <- readr::read_csv(csv_path, show_col_types = FALSE)
     }
 
     if (file.exists(raw_path)) {
-      ZB18a_raw <- readr::read_table(raw_path,
-                                     col_names = raw_col_names,
-                                     skip = 3, comment = "#",
-                                     show_col_types = FALSE)
+      if (astronomical_solution == "PT-ZB18a") {
+        raw <- readr::read_table(raw_path,
+                                 col_names = raw_col_names,
+                                 skip = 3, comment = "#",
+                                 show_col_types = FALSE)
+      } else {
+        raw <- readr::read_table(raw_path,
+                                 col_names = raw_col_names,
+                                 comment = "%",
+                                 show_col_types = FALSE)
+      }
     }
   } else {# files don't exist or force
-    if (!quiet) cli::cli_alert_info("The orbital solution ZB18a has not been downloaded.")
+    if (!quiet) cli::cli_alert_info("The astronomical solution {astronomical_solution} has not been downloaded.")
 
     # default to downloading/caching if not interactive (i.e. GitHub actions)
     if (force || !interactive()) {
@@ -104,31 +172,35 @@ get_ZB18a <- function(quiet = FALSE, force = FALSE) {
                               title = "Would you like to download and process it now?") == 1L
       if (download) {
         save_cache <- utils::menu(c("Yes", "No"),
-                                  title = "Would you like to save the results to {
-.file .csv} and {
-.file .rds}?") == 1L
+                                  title = "Would you like to save the results to .csv and .rds?") == 1L
       } else {
         save_cache <- FALSE
       }
     }
 
-    # the user would NOT like to download and process the orbital solution
+    # the user would NOT like to download and process the astronomical solution
     if (!download) {
-      cli::cli_abort("Cannot `get_ZB18a()` without downloading the file.")
+      cli::cli_abort("Cannot `get_ZB()` without downloading the file.")
     }
 
     # read the file from the website
-    ZB18a_raw <- readr::read_table(ZB18a_url,
-                                   col_names = raw_col_names,
-                                   skip = 3, comment = "#",
-                                   show_col_types = FALSE)
-    if (!quiet) cli::cli_alert_info("Read {.file {basename(raw_path)}} from website {.url {ZB18a_url}}.")
-
-    # calculate helper columns
-    ZB18a <- ZB18a_raw |> prepare_solution(quiet = quiet)
+    if (!quiet) cli::cli_alert_info("Reading {.file {basename(raw_path)}} from website {.url {url}}.")
+    if (astronomical_solution == "PT-ZB18a") {
+      raw <- readr::read_table(url,
+                               col_names = raw_col_names,
+                               skip = 3, comment = "#",
+                               show_col_types = FALSE)
+      # calculate helper columns
+      raw <- raw |> prepare_solution(quiet = quiet)
+    } else {
+      raw <- readr::read_table(url,
+                               comment = "%",
+                               col_names = raw_col_names,
+                               show_col_types = FALSE)
+    }
 
     if (!save_cache) {
-      return(ZB18a)
+      return(raw)
     } else {
       if (!dir.exists(cachedir)) {
         dir.create(cachedir, recursive = TRUE, showWarnings = TRUE)
@@ -137,20 +209,20 @@ get_ZB18a <- function(quiet = FALSE, force = FALSE) {
 
       # also copy the raw file to disk
       # even though we've read it in using read_table directly
-      curl::curl_download(ZB18a_url, destfile = raw_path)
+      curl::curl_download(url, destfile = raw_path)
       if (!quiet) cli::cli_alert_info("Saved {.file {basename(raw_path)}} to cache.")
 
       # write intermediate result to csv
-      readr::write_csv(ZB18a_raw, csv_path)
+      readr::write_csv(raw, csv_path)
       if (!quiet) cli::cli_alert_info("Saved cleaned-up {.file {basename(csv_path)}} to cache.")
 
       # write final result to rds cache
-      readr::write_rds(ZB18a, rds_path)
+      readr::write_rds(raw, rds_path)
       if (!quiet) {
         cli::cli_alert("Saved solution with helper columns {.file {basename(rds_path)}} to cache.",
                        "i" = "Future runs will read from the cache, unless you specify `force = TRUE`.")
       }
     }
-    return(ZB18a)
+    return(raw)
   }
 }