+++ R/snvec.R
@@ -1,7 +1,7 @@
 #' Calculate Earth’s Obliquity and Precession in the Past
 #'
 #' `snvec()` computes climatic precession and obliquity (or tilt) from an
-#' orbital solution (OS) input and input values for dynamical ellipticity
+#' astronomical solution (AS) input and input values for dynamical ellipticity
 #' (\eqn{E_{d}}{Ed}) and tidal dissipation (\eqn{T_{d}}{Td}). It solves a set
 #' of ordinary differential equations.
 #'
@@ -11,24 +11,29 @@
 #'   Defaults to `1.0`.
 #' @param td Tidal dissipation \eqn{T_{d}}{Td}, normalized to modern. Defaults
 #'   to `0.0`.
-# orbital_solution comes from get_solution
+# inherit astronomical_solution and quiet from get_solution
 #' @inheritParams get_solution
+#' @param os_ref_frame Character vector with the reference frame of the astronomical
+#'   solution. Either `"HCI"` for heliocentric inertial reference frame or
+#'   `"J2000"` for ecliptic J2000 reference frame. Defaults to `"HCI"` for
+#'   `HNBody` output.
+#' @param os_omt Longitude of ascending node of the solar equator relative to ECLIPJ2000.
+#' @param os_inct Inclination of the solar equator relative to ECLIPJ2000.
+# ODE solver parameters
 #' @param tres Output timestep resolution in thousands of years (kyr). Defaults
-#'   to `0.4`.
+#'   to `-0.4`. To determine the sign, think of from `0` to `tend` by timestep
+#'   `tres`.
 #' @param atol Numerical absolute tolerance passed to [deSolve::ode()]'s
 #'   `atol`. Defaults to `1e-5`.
 #' @param rtol Numerical relative tolerance passed to [deSolve::ode()]'s
 #'   `rtol`. Defaults to `0`.
 #' @param solver Character vector specifying the method passed to
-#'   [deSolve::ode()]'s `method`. Defaults to `"vode"` for stiff problems
-#'   with a variable timestep.
-# quiet comes from get_ZB18a. Force does too, but I hope that because we don't
-# use it here, it won't get inherited.
-#' @inheritParams get_ZB18a
+#'   [deSolve::ode()]'s `method`. Defaults to `"vode"` for stiff problems with
+#'   a variable timestep.
 #' @param output Character vector with name of desired output. One of:
 #'
 #'   * `"nice"` (the default) A [tibble][tibble::tibble-package] with the
-#'     columns `time`, `age`, `eei`, `epl`, `phi`, `cp`.
+#'     columns `time`, `t_ka`, `eei`, `epl`, `phi`, `cp`.
 #'
 #'   * `"full"` A [tibble][tibble::tibble-package] with all the computed and
 #'     interpolated columns.
@@ -39,9 +44,28 @@
 #'   information of Zeebe & Lourens (2022). The terms are explained in detail
 #'   in Zeebe (2022).
 #'
-#'   Note that the different ODE solver algorithm we use (Soetaert et al.,
-#'   2010) means that the R routine returns an evenly-spaced time grid, whereas
-#'   the C-routine has a variable time-step.
+#' @section Reference Frames of Astronomical Solutions:
+#'
+#' NASA provides their asteroid and planet positions in the ecliptic J2000
+#' reference frame, while long-term astronomical solution integrations are
+#' often performed in the heliocentric inertial reference frame (HCI) or in the
+#' inertial reference frame. This is to align the reference frame with the spin
+#' vector of the Sun, making J2 corrections intuitive to implement.
+#'
+#' Obliquity is typically given in the ecliptic reference frame, so snvec
+#' converts all outputs to J2000 if the `os_ref_frame` is equal to `"HCI"` and
+#' does no transformations if it is already in `"J2000"`.
+#'
+#' For this, it uses \eqn{\Omega_{\odot} = 75.5940}{OMT = 75.5940} and
+#' \eqn{i_{\odot = 7.155}}{INCT = 7.155} as in Zeebe (2017). You can overwrite
+#' these defaults with `os_omt` and `os_inct` if desired.
+#'
+#' @section ODE Solver:
+#'
+#' Note that the different ODE solver algorithm we use (Soetaert et al.,
+#' 2010) means that the R routine returns an evenly-spaced time grid, whereas
+#' the C-routine has a variable time-step.
+#' This means we need to explicitly step size (`tres`).
 #'
 #' @returns `snvec()` returns different output depending on the `outputs` argument.
 #'
@@ -50,11 +74,12 @@
 #'
 #'   * `time` Time \eqn{t} (days).
 #'
-#'   * `age` Age in thousands of years ago (ka).
+#'   * `t_ka` Time in thousands of years (ka).
 #'
-#'   * `eei` Orbital solution's eccentricity \eqn{e}, interpolated to output
-#'   timescale (-).
-#'
+# I removed this earlier
+# #'   * `eei` Astronomical solution's eccentricity \eqn{e}, interpolated to output
+# #'   timescale (-).
+# #'
 #'   * `epl` Calculated Obliquity \eqn{\epsilon} (radians).
 #'
 #'   * `phi` Calculated Precession \eqn{\phi} (radians) from ECLIPJ2000.
@@ -96,20 +121,20 @@
 #'
 #' If `output = "ode"`, it will return the raw output of the ODE solver, which
 #' is an object of class `deSolve` and `matrix`, with columns `time`, `sx`,
-#' `sy`, and `sz` (see above). This can be useful for i.e.
-#' [deSolve::diagnostics()].
+#' `sy`, and `sz`. This can be useful for i.e. [deSolve::diagnostics()].
 #'
 #' @seealso
 #'
 #' * [deSolve::ode()] from Soetaert et al., (2010) for the ODE solver that we
 #'   use.
 #'
-#' * [get_ZB18a()] Documents the default orbital solution input.
-#'
-#' * [get_solution()] A general function that in the future may be used to get
-#'   other orbital solutions.
+#' * [get_solution()] Get astronomical solutions.
 #'
 #' @references
+#'
+#' Zeebe, R.E. (2017). Numerical Solutions for the Orbital Motion of the Solar
+#'   System over the Past 100 Myr: Limits and New Results. _The Astronomical
+#'   Journal_, 154(5), \doi{10.3847/1538-3881/aa8cce}.
 #'
 #' Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
 #'  Paleocene–Eocene boundary age constrained by geology and astronomy.
@@ -136,15 +161,17 @@
 #' # default call
 #' \donttest{
 #' snvec()
-#' # remove the directory with the cached orbital solution to clean up
+#' # remove the directory with the cached astronomical solution to clean up
 #' unlink(tools::R_user_dir("snvecR", which = "cache"), recursive = TRUE)
 #' }
 #' @export
 snvec <- function(tend = -1e3,
                   ed = 1,
                   td = 0,
-                  orbital_solution = "ZB18a",
-                  tres = 0.4,
+                  astronomical_solution = "PT-ZB18a",
+                  os_ref_frame = "HCI",
+                  os_omt = NULL, os_inct = NULL,
+                  tres = -0.4,
                   atol = 1e-5,
                   rtol = 0,
                   solver = "vode",
@@ -156,30 +183,22 @@ snvec <- function(tend = -1e3,
     cli::cli_abort(c("{.var output} must be one of {.or {.q {outputs}}}.",
                      "x" = "You've supplied {.q {output}}."))
   }
-  ## tend
-  if (tend >= 0) {
-    cli::cli_abort(c("{.var tend} must be < 0.",
-      "x" = "You've supplied {tend}."
-    ))
-  }
-  if (tend < -1e5) {
-    cli::cli_abort(c("{.var tend} must be > the orbital solution {-1e5}, e.g. -100 Ma.",
-      "x" = "You've supplied {tend}."
-    ))
-  }
 
   ## tres
-  ## a quick dumb input test for now
-  if (tres < tend) {
-    cli::cli_abort(c("{.var tres} must be < {.var tend}.",
-      "i" = "{.var tres} = {tres}",
-      "i" = "{.var tend} = {tend}"
-    ))
+  if (sign(tres) != sign(tend)) {
+    cli::cli_abort(c("{.var tres} must be given in the same sign as {.var tend}.",
+                     "i" = "{.var tres} = {tres}",
+                     "i" = "{.var tend} = {tend}"
+                     ))
   }
 
-  if (tres <= 0) {
-    cli::cli_abort(c("{.var tres} must be > 0.",
-                     "i" = "{.var tres} = {tres}"))
+
+  ## a quick dumb input test for now
+  if (abs(tres) > abs(tend)) {
+    cli::cli_abort(c("abs({.var tres}) must be < abs({.var tend}).",
+                     "i" = "{.var tres} = {tres}",
+                     "i" = "{.var tend} = {tend}"
+                     ))
   }
 
   # this warning is too strict and kind of annoying
@@ -196,6 +215,50 @@ snvec <- function(tend = -1e3,
                     "*" = "See Zeebe & Lourens 2022 Pal&Pal <https://doi.org/10.1029/2021PA004349>."))
   }
 
+  if (!"data.frame" %in% class(astronomical_solution) &&
+        !grepl("^PT-", astronomical_solution)) {
+    cli::cli_abort(c("Astronomical Solution must contain all orbital parameters",
+                     "i" = "Did you mean to specify {.q PT-ZB18a}?"))
+  }
+
+  hci_refs <- c("heliocentric intertial", "HCI")
+  j2000_refs <- c("ecliptic", "ECLIPJ2000", "J2000")
+  if (!os_ref_frame %in% c(hci_refs, j2000_refs)) {
+    cli::cli_abort(c(
+      "i" = "{.var os_ref_frame} must be one of {.or {.q {reference_frames}}}",
+      "x" = "{.var os_ref_frame} = {.q {os_ref_frame}}"
+    ))
+  } else {
+    # get rid of aliases
+    if (os_ref_frame %in% hci_refs) os_ref_frame <- "HCI"
+    if (os_ref_frame %in% j2000_refs) os_ref_frame <- "J2000"
+  }
+
+  # if the user specifies a value for os_omt and/or os_inct, they cannot
+  # specify a non-default os_ref_frame
+  if (os_ref_frame != "HCI") {
+    if (!is.null(os_omt)) {
+      cli::cli_abort("Specified both {.var os_ref_frame} and {.var os_omt}.")
+    }
+    if (!is.null(os_inct)) {
+      cli::cli_abort("Specified both {.var os_ref_frame} and {.var os_inct}.")
+    }
+  }
+
+  # if the reference frame is already J2000, just set the angles to 0
+  # (this makes the codebase simpler, no need for ifelse statements throughout.)
+  # but we will put in explicit ifelse's for optimisation
+  if (os_ref_frame == "HCI" && is.null(os_omt) && is.null(os_inct)) {
+    OMT <- 75.5940
+    INCT <- 7.155
+  } else if (os_ref_frame == "J2000") {
+    if (is.null(os_omt)) OMT <- 0.
+    if (is.null(os_inct)) INCT <- 0.
+  }
+  # or if the user has specified their own rotation, that's fine too!
+  if (!is.null(os_omt)) OMT <- os_omt
+  if (!is.null(os_inct)) INCT <- os_inct
+
   if (atol < 1e-12 | atol > 1e-3) {
     cli::cli_warn("Input absolute tolerance should be between 1e-12 and 1e-3.")
   }
@@ -203,24 +266,35 @@ snvec <- function(tend = -1e3,
     cli::cli_warn("Input relative tolerance should be smaller than 1e-3.")
   }
 
-  dat <- get_solution(orbital_solution = orbital_solution)
+  dat <- get_solution(astronomical_solution = astronomical_solution, quiet = quiet)
+
+  if ((sign(tend) != sign(dat$t_ka[2])) || (abs(tend) > max(abs(dat$t_ka)))) {
+    cli::cli_abort(c("{.var tend} must fall within astronomical solution age.",
+                     "i" = "The astronomical solution {sign(dat$t_ka[2])*max(abs(dat$t_ka))}.",
+                     "x" = "{.var tend} = {tend}."
+                     ))
+  }
 
   # message user about inputs
   if (!quiet) {
     startdate <- lubridate::now()
     cli::cli_inform(c(
-      ## "This is {VER}",
-      ## "Richard E. Zeebe",
-      ## "Ilja J. Kocken",
+      "This is {VER}",
+      "Richard E. Zeebe",
+      "Ilja J. Kocken",
+      "",
       "Integration parameters:",
-      "*" = "{.var tend} = {tend} ka",
-      "*" = "{.var ed} = {ed}",
-      "*" = "{.var td} = {td}",
-      "*" = "{.var orbital_solution} = {.q {orbital_solution}}",
-      "*" = "{.var tres} = {tres} kyr",
-      "*" = "{.var atol} = {atol}",
-      "*" = "{.var rtol} = {rtol}",
-      "*" = "{.var solver} = {.q {solver}}",
+      "*" = "{.var tend} = {.val {tend}} ka",
+      "*" = "{.var ed} = {.val {ed}}",
+      "*" = "{.var td} = {.val {td}}",
+      "*" = "{.var astronomical_solution} = {.val {if ('data.frame' %in% class(astronomical_solution)) 'user provided' else astronomical_solution}}",
+      "*" = "{.var os_ref_frame} = {.val {os_ref_frame}}",
+      "*" = "{.var os_omt} = {if (is.null(os_omt)) 'defaulting to' else ''} {.val {OMT}}",
+      "*" = "{.var os_inct} = {if (is.null(os_inct)) 'defaulting to' else ''} {.val {INCT}}",
+      "*" = "{.var tres} = {.val {tres}} kyr",
+      "*" = "{.var atol} = {.val {atol}}",
+      "*" = "{.var rtol} = {.val {rtol}}",
+      "*" = "{.var solver} = {.val {solver}}",
       "i" = "started at {.q {startdate}}"
     ))
   }
@@ -253,8 +327,12 @@ snvec <- function(tend = -1e3,
     as.matrix() |>
     as.vector()
 
-  # transform n => n'
-  np <- euler(ninit, inct, omt, TRUE)
+  if (os_ref_frame != "J2000") {
+    # transform n => n'
+    np <- euler(ninit, inct, omt, TRUE)
+  } else {
+    np <- ninit
+  }
 
   # solve quadratic equation for s0'y
   a <- np[2] * np[2] + np[3] * np[3]
@@ -268,8 +346,12 @@ snvec <- function(tend = -1e3,
   s0p[1] <- 0
   as.matrix(s0p)
 
-  # transform s0' to s0
-  s0 <- euler(s0p, inct, omt, 0)
+  if (os_ref_frame != "J2000") {
+    # transform s0' to s0
+    s0 <- euler(s0p, inct, omt, FALSE)
+  } else {
+    s0 <- s0p
+  }
 
   ## set the deSolve state
   state <- c(
@@ -280,8 +362,6 @@ snvec <- function(tend = -1e3,
 
   ## define deSolve parameters
   parameters <- c(
-    ed = ed,
-    td = td,
     k0d = k0d,
     wdw = wdw,
     ndn = ndn
@@ -292,12 +372,12 @@ snvec <- function(tend = -1e3,
   # derivatives. RHS of DEQs for spin vector s = y
   eqns <- function(t, state, parameters) {
     with(as.list(c(state, parameters)), {
-      # the the index of the current timestep in the orbital solution
+      # the the index of the current timestep in the astronomical solution
       m <- min(round(abs(t / dts) + 1), nrow(dat))
-
-      ## quickly interpolate orbital solution input value at the current
-      ## timestep
+      # quickly interpolate astronomical solution input value at the current
+      # timestep
       dx <- t - dat$t[m]
+
       qqi <- qinterp(dat$qq, dts, dx, m)
       ppi <- qinterp(dat$pp, dts, dx, m)
       cci <- qinterp(dat$cc, dts, dx, m)
@@ -342,10 +422,10 @@ snvec <- function(tend = -1e3,
 
   ## a linear sequence of steps
   times <- seq(0, tend * KY2D,
-               by = -tres * KY2D
+               by = tres * KY2D
                # ~ tres = 0.4 is the average diff in the C-output
                # snv_sout$time |> diff() |> median() = 0.396013
-  )
+               )
 
   ## solve it
   ## [[file:snvec-3.7.5/snvec-3.7.5.c::%%% solver][odeint()]]
@@ -361,7 +441,7 @@ snvec <- function(tend = -1e3,
   )
 
   ## print the final values for s
-  ## This is at t = tend, it's going back from 0 to -time
+  ## This is at t = tend, it's going back from 0 to -time by default.
   fin <- out[nrow(out), ]
   u <- as.vector(c(fin[2], fin[3], fin[4]))
 
@@ -373,9 +453,9 @@ snvec <- function(tend = -1e3,
   if (!quiet) {
     cli::cli_inform(c(
       "Final values:",
-      "*" = "s[1][2][3]: {paste(fin[2], fin[3], fin[4])}",
+      "*" = "s[1][2][3]: {.val {c(fin[2], fin[3], fin[4])}}",
       ## "*" = "s-error = |s|-1: {sqrt(pracma::dot(u, u))-1}"
-      "*" = "s-error = |s|-1: {sqrt(u[1]^+u[2]^2+u[3]^2)-1}"
+      "*" = "s-error = |s|-1: {.val {sqrt(u[1]^2+u[2]^2+u[3]^2)-1}}"
     ))
   }
 
@@ -384,11 +464,11 @@ snvec <- function(tend = -1e3,
     return(out)
   }
 
-  ## interpolate the full orbital solution onto output timescale
+  ## interpolate the full astronomical solution onto output timescale
   fin <- out |>
     tibble::as_tibble() |>
     dplyr::mutate(
-      age = -.data$time / KY2D,
+      t_ka = .data$time / KY2D,
       nnx = approxdat(dat, "nnx")(.data$time),
       nny = approxdat(dat, "nny")(.data$time),
       nnz = approxdat(dat, "nnz")(.data$time),
@@ -398,50 +478,55 @@ snvec <- function(tend = -1e3,
       lani = approxdat(dat, "lanu")(.data$time)
     )
 
-  ## calculate obliquity
-  fin <- fin |>
-    # calculate the dotproduct, Richard's vvdot
-    dplyr::mutate(
-      tmp = .data$sx * .data$nnx + .data$sy * .data$nny + .data$sz * .data$nnz,
-      epl = acos(.data$tmp)
-    )
+  ## ## calculate obliquity
+  ## fin <- fin |>
+  ##   # calculate the dotproduct, Richard's vvdot
+  ##   dplyr::mutate(
+  ##     tmp = .data$sx * .data$nnx + .data$sy * .data$nny + .data$sz * .data$nnz,
+  ##     epl = acos(.data$tmp)
+  ##   )
 
-  ## calculate precession and climatic precession
   fin <- fin |>
     # for each row, NOTE this makes it very slow!!
     dplyr::rowwise() |>
-    # create list columns of matrices
     # extract sx, sy, sz, and nnx, nny, nnz
     dplyr::mutate(
+      # create list columns of vectors
       u = list(matrix(c(.data$sx, .data$sy, .data$sz), ncol = 1, nrow = 3)),
       nv = list(matrix(c(.data$nnx, .data$nny, .data$nnz), ncol = 1, nrow = 3)),
-      # coords: fixed HCI => moving orbit plane
-      up = list(euler(.data$u, .data$inci / R2D, .data$lani / R2D, 0)),
+
+      # calculate obliquity
+      tmp = pracma::dot(.data$u, .data$nv),
+      ## tmp = .data$sx * .data$nnx + .data$sy * .data$nny + .data$sz * .data$nnz,
+      # calculate obliquity
+      epl = acos(.data$tmp),
+
+      # coords: fixed => moving orbit plane
+      up = list(euler(.data$u, .data$inci / R2D, .data$lani / R2D, FALSE)),
       # coords: relative to phi(t=0)=0 at J2000
-      up = list(euler(.data$up, 0, -(.data$lani + OMT) / R2D - pi / 2, 0)),
+      up = list(euler(.data$up, 0, -(.data$lani + OMT) / R2D + pi / 2, FALSE)),
       # get 2nd and 1st column of up
+
+      ## calculate axial precession
       phi = map2_dbl(.data$up[2, ], .data$up[1, ], atan2)
     ) |>
-    dplyr::ungroup() |>
+    dplyr::ungroup() |> # end rowwise
     dplyr::mutate(
       # normalize to first value of phi
-      tmp = first(.data$phi),
-      phi = .data$phi - .data$tmp,
-      cp = .data$eei * sin((.data$lphi + OMT) / R2D - .data$phi),
-      # phi is now mapped between 0 and 2 pi, whereas RZ's output is wrapped
-      # between -pi and pi
-      phi = ifelse(.data$phi > pi, .data$phi - 2 * pi, .data$phi)
+      phi = .data$phi - first(.data$phi),
+      # calculate climatic precession
+      cp = .data$eei * sin((.data$lphi + OMT) / R2D - .data$phi)
     )
 
   ## message user about final values
   if (!quiet) {
     cli::cli_inform(
       c("Final values:",
-        "*" = "obliquity: {fin[nrow(fin), 'epl']} rad",
-        "*" = "precession: {fin[nrow(fin), 'phi']} rad",
+        "*" = "obliquity: {.val {fin[nrow(fin), 'epl']}} rad",
+        "*" = "precession: {.val {fin[nrow(fin), 'phi']}} rad",
         "i" = "stopped at {.q {lubridate::now()}}",
-        "i" = "total duration:  {lubridate::as.duration(round(lubridate::now() - startdate, 2))}"
-      )
+        "i" = "total duration: {.val {lubridate::as.duration(round(lubridate::now() - startdate, 2))}}"
+        )
     )
   }
 
@@ -450,7 +535,7 @@ snvec <- function(tend = -1e3,
     # we transform the deSolve parameters into simple numeric columns
     # this is so they work better with things like bind_rows etc. via vctrs
     dplyr::mutate(dplyr::across(
-      tidyselect::all_of(c("time", "sx", "sy", "sz", "age", "epl")),
+      tidyselect::all_of(c("time", "t_ka", "sx", "sy", "sz", "epl")),
       as.numeric))
 
   if (output == "all") {
@@ -459,12 +544,9 @@ snvec <- function(tend = -1e3,
 
   if (output == "nice") {
     fin |>
-      # get rid of columns that we do not like
-      dplyr::select(-tidyselect::all_of(c(
-        "sx", "sy", "sz",
-        "nnx", "nny", "nnz",
-        "inci", "lphi", "lani",
-        "tmp",
-        "u", "nv", "up")))
+      dplyr::select(tidyselect::all_of(c(
+        "time", "t_ka",
+        "epl", "phi", "cp"
+      )))
   }
 }